# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SORRISO PERFEITO API",
    "description": "API interativa para gestão de clínica odontológica, incluindo controle de acessos (usuários), pacientes, prontuário (anamnese), odontólogos, agendamento de consultas e catalogação de procedimentos clínicos.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Servidor de Desenvolvimento Local"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/auth/login": {
      "post": {
        "security": [],
        "tags": [
          "Autenticação"
        ],
        "summary": "Realizar login no sistema",
        "description": "Valida as credenciais e retorna um Token JWT para autenticação.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "senha"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "test_unique@example.com"
                  },
                  "senha": {
                    "type": "string",
                    "format": "password",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEnvelope"
                },
                "example": {
                  "success": true,
                  "data": {
                    "message": "Login realizado com sucesso",
                    "token": "ey...token_de_sessao",
                    "usuario": {
                      "id": 10,
                      "nome": "Test User",
                      "email": "test_unique@example.com",
                      "tipo": "PACIENTE"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados ausentes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/recuperar-senha": {
      "post": {
        "security": [],
        "tags": [
          "Autenticação"
        ],
        "summary": "Solicitar recuperação de senha",
        "description": "Envia um token de recuperação via e-mail se o cadastro existir.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Instruções enviadas para redefinir senha",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/auth/redefinir-senha": {
      "post": {
        "security": [],
        "tags": [
          "Autenticação"
        ],
        "summary": "Redefinir senha usando Token",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "token",
                  "senha",
                  "senha_repeat"
                ],
                "properties": {
                  "token": {
                    "type": "string",
                    "description": "Token enviado no e-mail de recuperação"
                  },
                  "senha": {
                    "type": "string",
                    "minLength": 6,
                    "example": "newpassword123"
                  },
                  "senha_repeat": {
                    "type": "string",
                    "example": "newpassword123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Senha alterada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação ou token inválido/expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponseEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Operação executada com sucesso"
              }
            }
          }
        }
      },
      "ErrorResponseEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Descrição detalhada do erro ou falha de validação"
          }
        }
      },
...
    }
  }
}